Extract the ObjectMapperCustomizer logic from DefaultSchemaGenerator
- because ObjectMapper is initialized during STATIC_INIT which may cause
troubles when a SchemaGeneratorConfigCustomizer depends on runtime
configuration mapping
